actualizacion de archivos

+++ docs/scripts/02_carretillas.sql
@@ -26,7 +26,7 @@ CREATE TABLE
 CREATE TABLE
     `carretillaanon` (
         `anoncod` varchar(128) NOT NULL,
-        `productId` bigint(18) NOT NULL,
+        `productId` int(11) NOT NULL,
         `crrctd` int(5) NOT NULL,
         `crrprc` decimal(12, 2) NOT NULL,
         `crrfching` datetime NOT NULL,
